Move AsyncLock
Don't open the browser on npm run serve

Add start script that does open the browser
