+++ packages/dev/core/src/Misc/index.ts
@@ -67,3 +67,4 @@ export * from "./dumpTools";
 export * from "./greasedLineTools";
 export * from "./equirectangularCapture";
 export * from "./decorators.serialization";
+export * from "./asyncLock";
