+++ packages/tools/viewer-alpha/package.json
@@ -6,7 +6,8 @@
     "description": "A viewer using BabylonJS to display 3D elements natively",
     "scripts": {
         "preserve": "npm run build",
-        "serve": "webpack serve --open",
+        "serve": "webpack serve",
+        "start": "npm run serve -- --open",
         "build": "npm run compile",
         "clean": "rimraf dist && rimraf *.tsbuildinfo",
         "compile": "tsc -b tsconfig.build.json",
